Refactor : passinger profile edit page api 관련 파일 > member 컴포넌트와 분리
