--- src/app/(client)/apis/getMember/types.ts
@@ -1,20 +1,0 @@
-export type PostUploadParams = {
-  file: string;
-};
-
-export type UploadDto = {
-  data: string;
-  message: string;
-  status: string;
-};
-
-export type PutProfileEditParams = {
-  nickname: string;
-  profileImage: string;
-};
-
-export type ProfileEditDto = {
-  message: string;
-  status: string;
-  data: PutProfileEditParams;
-};
