+++ src/app/(client)/profile/apis/putProfileEdit.ts
@@ -2,8 +2,8 @@ import { API_ROUTES } from '@/constants/routes';
 
 import { fetchPut } from '@/apis/fetch';
 
-import { PutProfileEditParams } from '@/app/(client)/apis/getMember/types';
 import { ApiSuccessResponse } from '@/apis/types';
+import { PutProfileEditParams } from '@/app/(client)/profile/apis/types';
 
 const putProfileEdit = async (params: PutProfileEditParams) => {
   const response = await fetchPut<
