+++ src/app/(client)/profile/apis/types.ts
@@ -1,1 +1,12 @@
 export type GetProfileListType = 'history' | 'driveHis';
+
+export type PutProfileEditParams = {
+  nickname: string;
+  profileImage: string;
+};
+
+export type ProfileEditDto = {
+  message: string;
+  status: string;
+  data: PutProfileEditParams;
+};
