+++ src/app/(client)/profile/apis/usePutProfileEdit.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@tanstack/react-query';
 
-import putProfileEdit from '@/app/(client)/apis/getMember/putProfileEdit';
+import putProfileEdit from '@/app/(client)/profile/apis/putProfileEdit';
 
 const usePutProfileEdit = () => {
   return useMutation({
